zalcano. Moved code onto another overlay and gave above widgets rights.
