+++ runelite-client/src/main/java/net/runelite/client/plugins/zalcano/ZalcanoOverlay.java
@@ -31,8 +31,6 @@ import java.awt.Dimension;
 import java.awt.Graphics2D;
 import java.awt.Polygon;
 import java.awt.Shape;
-import java.awt.Rectangle;
-import java.awt.Font;
 import java.util.List;
 import javax.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
@@ -41,7 +39,6 @@ import net.runelite.api.Client;
 import net.runelite.api.GameObject;
 import net.runelite.api.GraphicsObject;
 import net.runelite.api.Perspective;
-import net.runelite.api.Point;
 import net.runelite.api.coords.LocalPoint;
 import net.runelite.api.coords.WorldPoint;
 import net.runelite.client.ui.overlay.Overlay;
@@ -83,7 +80,7 @@ public class ZalcanoOverlay extends Overlay
 		this.util = util;
 		this.client = client;
 
-		setLayer(OverlayLayer.ABOVE_WIDGETS);
+		setLayer(OverlayLayer.ABOVE_SCENE);
 		setPosition(OverlayPosition.DYNAMIC);
 		setPriority(OverlayPriority.HIGH);
 	}
@@ -115,14 +112,6 @@ public class ZalcanoOverlay extends Overlay
 		if (config.highlightZalcanoHull())
 		{
 			renderZalcano(graphics);
-		}
-		if (config.playerCountAtZalcano())
-		{
-			renderPlayerCount(graphics);
-		}
-		if (config.highlightMiningSpotAtMap())
-		{
-			renderRockToMineAtMap(graphics);
 		}
 
 		//has their own configs within this method
@@ -177,25 +166,6 @@ public class ZalcanoOverlay extends Overlay
 				OverlayUtil.renderPolygon(graphics, poly, !util.projectileExists() ? green : Color.RED);
 				OverlayUtil.renderTextLocation(graphics, glowingRock.getCanvasLocation(), !util.projectileExists() ? ZalcanoUtil.mine : ZalcanoUtil.warning, !util.projectileExists() ? green : Color.RED);
 			}
-		}
-	}
-
-	private void renderRockToMineAtMap(Graphics2D graphics)
-	{
-		GameObject glowingRock = util.getGlowingRock();
-
-		if (glowingRock != null)
-		{
-			Point rockPoint = glowingRock.getMinimapLocation();
-			OverlayUtil.renderMinimapLocation(graphics, rockPoint, Color.BLUE);
-			Point rockPoint1 = new Point (rockPoint.getX() + 4, rockPoint.getY());
-			OverlayUtil.renderMinimapLocation(graphics, rockPoint1, Color.MAGENTA);
-			Point rockPoint2 = new Point (rockPoint.getX() - 4, rockPoint.getY());
-			OverlayUtil.renderMinimapLocation(graphics, rockPoint2, Color.MAGENTA);
-			Point rockPoint3 = new Point (rockPoint.getX(), rockPoint.getY() + 4);
-			OverlayUtil.renderMinimapLocation(graphics, rockPoint3, Color.MAGENTA);
-			Point rockPoint4 = new Point (rockPoint.getX(), rockPoint.getY() - 4);
-			OverlayUtil.renderMinimapLocation(graphics, rockPoint4, Color.MAGENTA);
 		}
 	}
 
@@ -265,18 +235,6 @@ public class ZalcanoOverlay extends Overlay
 			OverlayUtil.renderPolygon(graphics, poly, color);
 			OverlayUtil.renderTextLocation(graphics, plugin.getZalcano().getCanvasTextLocation(graphics, text, plugin.getZalcano().getLogicalHeight() / 2), text, color);
 		}
-	}
-
-	private void renderPlayerCount(Graphics2D graphics)
-	{
-		Rectangle rect = new Rectangle(client.getCanvasWidth() / 2 - 15, 0, 40, 25);
-		int[] xpoints = {rect.x, rect.x + rect.width, rect.x + rect.width, rect.x};
-		int[] ypoints = {rect.y, rect.y, rect.y + rect.height, rect.y + rect.height};
-
-		graphics.setFont(new Font("Comic Sans", Font.BOLD, 20));
-		OverlayUtil.renderFilledPolygon(graphics, rect, Color.BLACK);
-		String playerCount = java.lang.Integer.toString(client.getPlayers().size());
-		OverlayUtil.renderTextLocation( graphics, new Point(client.getCanvasWidth() / 2, 20), playerCount, Color.WHITE);
 	}
